tests: Fix and simplify restoring state after auth tests
There was a bug with  0ee9444a2646d0445224246707ebac61a2f9b545 which caused
tests to fail when process.env.STRAVA_ACCESS_TOKEN was NOT set in the
environment. This issue is not triggered in the Travis CI environment because
the variable IS there.

The problem is triggered by logic like this:

    process.env.STRAVA_ACCESS_TOKEN = originalToken;

As documented in the Node.js `process` docs, assigning a value to process.env
always casts it to a string. So when original token is undefined it restored as
the string "undefined".... which is not undefined.

Since the environment variable is now defined as "undefined", this will
override any correct values found in the config file.

The `env-restorer` module was created to solve exactly this problem and was added as
a devDependency.

Using Mocha's afterEach() cleans up the test syntax.