tests: fail gracefully if test account does not have expected data.  Fixes #53
The test suite generally assumed it a "happy path". It did not check if errors
were returned from API calls or if empty or missing values were returned
when fetching sample data.

This update adds a lot of error/reality checking to the test suite and
returns better diagnostics if expected data is not there.

It also increases the chance that the segment-related tests will pass by
finding a "sample activity" that contains an achievement (which are
often segments-- not perfect, but better than just assuming the most
recent activity will work!)