BugFix: For 'done is not a function' with args fail validation.  (breaking change)
Note this is a breaking behavior change.

Before these programming errors were handled in error backs, now they
are thrown exceptions.

The option was chosen because exceptions work for both original code
path and the promise code path in a unified way.

Related refactoring was done to DRY the the validation code in
./lib/activities.js