Bugfix for when there is no response
Seems that the Strava API has some weird behavior where the API is not actually down, just responding without a response causing the below: 
TypeError: Cannot read property 'headers' of undefined at line 220. Haven't dug to deep in your code so you might want to fix it in another way but this should at least guard against that.

My app goes down maybe once a week or something like that because of this.